add: add single file for all get put post delete request and add a separate file for common type and interface
