+++ src/SignUp/commonTypeInterface.ts
@@ -1,0 +1,15 @@
+export interface Product {
+  id: string;
+  name: string;
+  price: number;
+  image: string;
+  userId: string;
+}
+export interface User {
+  userId: string;
+  fullName: string;
+  email: string;
+  password: string;
+  role: "buyer" | "seller";
+  cart: Product[];
+}
